+++ CHANGES.rst
@@ -21,6 +21,8 @@ Changelog
 - requires kasperl>=0.0.2 now
 - added plugin hierarchy for `Phonemizer` plugins and `phonemize` filter that applies a phonemizer
   to the text of speech data
+- added plugin hierarchy for `Watermarker` plugins and `watermark` filter that applies a watermark algorithm
+  to the audio data passing through
 
 
 0.1.0 (2025-10-31)